+++ .github/workflows/scripts/generate-srcinfo.sh
@@ -1,4 +1,0 @@
-#!/usr/bin/env bash
-set -euo pipefail
-
-makepkg --printsrcinfo > .SRCINFO